+++ PermissionServer/ITokenStore.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace PermissionServer
 {
     /// <summary>
@@ -11,17 +13,20 @@ namespace PermissionServer
     {
         /// <summary>
         /// Adds a new time-limited token.
-        /// If one already exists it will be replaced (with a new lifetime).
+        /// If the `MaximumActivePerKey` is reached no more tokens are issued
+        /// and the return value is null. The user will need to wait for the
+        /// oldest active key to expire before another will be allowed.
         /// </summary>
         /// <param name="key">Email address or other unique identifier.</param>
-        Token Add(string key);
+        Token? Add(string key);
 
-        /// <summary>Finds an existing token.</summary>
+        /// <summary>Finds an existing (unexpired) token.</summary>
         /// <param name="key">Email address or other unique identifier.</param>
-        Token? Get(string key);
+        /// <param name="confirmationCode">The related confirmation code.</param>
+        Token? Get(string key, string confirmationCode);
 
         /// <summary>Remove any existing matching token.</summary>
-        void Remove(string key);
+        void Remove(string key, string confirmationCode);
 
         /// <summary>
         /// Removes all expired tokens (frees resources).