+++ PermissionServer/PermissionServer.cs
@@ -20,10 +20,12 @@ namespace PermissionServer
         }
 
         /// <summary>Generates a new limited-lifetime token and emails it.</summary>
-        public async Task StartConfirmation(string emailAddress, string confirmationUrl)
+        public async Task<bool> StartConfirmation(string emailAddress, string confirmationUrl)
         {
             var token = tokenStore.Add(emailAddress);
+            if (token == null) return false;
             await mailer.SendConfirmation(emailAddress, token, confirmationUrl);
+            return true;
         }
 
         /// <summary>
@@ -31,15 +33,12 @@ namespace PermissionServer
         /// address (and has not expired). If it doesn't then the
         /// return value states why.
         /// </summary>
-        public ConfirmationStatus CompleteConfirmation(string emailAddress, string confirmationCode)
+        public bool CompleteConfirmation(string emailAddress, string confirmationCode)
         {
-            var token = tokenStore.Get(emailAddress);
-            if (token == null) return ConfirmationStatus.NoTokenFound;
-            if (token.IsExpired) return ConfirmationStatus.HasExpired;
-            if (token.DoesNotMatch(confirmationCode)) return ConfirmationStatus.DoesNotMatchToken;
-
-            if (options.Tokens.SingleUse) tokenStore.Remove(emailAddress);
-            return ConfirmationStatus.Okay;
+            var token = tokenStore.Get(emailAddress, confirmationCode);
+            if (token == null) return false;
+            if (options.Tokens.SingleUse) tokenStore.Remove(emailAddress, confirmationCode);
+            return true;
         }
     }
 }